Limit use of chaincode/module with binary launcher
The module path should be exercised for package and install but is not
needed when building chaincode with the external binary builder.

- Replace most use of chaincode/module with chaincode/simple
- Move signal handling from chaincode/module to chaincode/simple
