+++ integration/lifecycle/interop_test.go
@@ -114,9 +114,9 @@ var _ = Describe("Release interoperability", func() {
 		chaincode = nwo.Chaincode{
 			Name:            "mycc",
 			Version:         "0.0",
-			Path:            components.Build("github.com/hyperledger/fabric/integration/chaincode/module"),
+			Path:            components.Build("github.com/hyperledger/fabric/integration/chaincode/simple/cmd"),
 			Lang:            "binary",
-			PackageFile:     filepath.Join(testDir, "modulecc.tar.gz"),
+			PackageFile:     filepath.Join(testDir, "simplecc.tar.gz"),
 			SignaturePolicy: `OR ('Org1MSP.member','Org2MSP.member')`,
 			Sequence:        "1",
 			InitRequired:    false,
@@ -199,9 +199,9 @@ var _ = Describe("Release interoperability", func() {
 			chaincode = nwo.Chaincode{
 				Name:            "mycc",
 				Version:         "0.0",
-				Path:            components.Build("github.com/hyperledger/fabric/integration/chaincode/module"),
+				Path:            components.Build("github.com/hyperledger/fabric/integration/chaincode/simple/cmd"),
 				Lang:            "binary",
-				PackageFile:     filepath.Join(testDir, "modulecc.tar.gz"),
+				PackageFile:     filepath.Join(testDir, "simplecc.tar.gz"),
 				SignaturePolicy: `OR ('Org1MSP.member','Org2MSP.member')`,
 				Sequence:        "1",
 				InitRequired:    false,
@@ -225,9 +225,9 @@ var _ = Describe("Release interoperability", func() {
 			chaincode := nwo.Chaincode{
 				Name:            "mycc",
 				Version:         "0.0",
-				Path:            components.Build("github.com/hyperledger/fabric/integration/chaincode/module"),
+				Path:            components.Build("github.com/hyperledger/fabric/integration/chaincode/simple/cmd"),
 				Lang:            "binary",
-				PackageFile:     filepath.Join(testDir, "modulecc.tar.gz"),
+				PackageFile:     filepath.Join(testDir, "simplecc.tar.gz"),
 				SignaturePolicy: `AND ('Org1MSP.member','Org2MSP.member')`,
 				Sequence:        "1",
 				InitRequired:    true,
